fix age int byte

+++ backend/app/auth/auth_model.go
@@ -1,11 +1,11 @@
 package auth
 
 type SignUpReq struct {
-	IdToken string `json:"idToken"`
-	Social      string `json:"social"  example:"google"`
-	Nickname    string `json:"nickname"`
-	Gender      string `json:"gender" example:"m"`
-	Age         uint8  `json:"age" example:"20"`
+	IdToken  string `json:"idToken"`
+	Social   string `json:"social"  example:"google"`
+	Nickname string `json:"nickname"`
+	Gender   string `json:"gender" example:"m"`
+	Age      uint32 `json:"age" example:"20"`
 }
 
 type SignUpInfo struct {
@@ -13,6 +13,6 @@ type SignUpInfo struct {
 	Social     string `json:"social"  example:"google"`
 	Nickname   string `json:"nickname"`
 	Gender     string `json:"Gender" example:"m"`
-	Age        uint8  `json:"age" example:"20"`
+	Age        uint32 `json:"age" example:"20"`
 	SignUpDate string `json:"sign_up_date"`
 }
